relion_reconstruct: repair --ewald --newbox.
Old codes for --ewald --newbox cropped particles in the real space after
applying CTFP or Q but have been broken (probably) since
the introduction of optics groups in RELION 3.1.

Now, box size differences between the input particles
(read from an optics table) and the output box (requested by --newbox)
are handled in BackProjector by applying a scaling matrix to
the rotation matrix with obsModel.applyScaleDifference().
This works with/without --ewald.

Thanks to Martin Obr for the bug report and testing.